restructure everything
this doesn't actually build yet, i marked a lot of directories as "excluded" for now
