+++ http/src/main/java/net/daporkchop/lib/http/common/header/HeaderMap.java
@@ -13,7 +13,7 @@
  *
  */
 
-package net.daporkchop.lib.http.util.header;
+package net.daporkchop.lib.http.common.header;
 
 import lombok.NonNull;
