--- http/src/main/java/net/daporkchop/lib/http/util/header/Header.java
@@ -1,97 +1,0 @@
-/*
- * Adapted from the Wizardry License
- *
- * Copyright (c) 2018-2019 DaPorkchop_ and contributors
- *
- * Permission is hereby granted to any persons and/or organizations using this software to copy, modify, merge, publish, and distribute it. Said persons and/or organizations are not allowed to use the software or any derivatives of the work for commercial use or any other means to generate income, nor are they allowed to claim this software as their own.
- *
- * The persons and/or organizations are also disallowed from sub-licensing and/or trademarking this software without explicit permission from DaPorkchop_.
- *
- * Any persons and/or organizations using this software must disclose their source code and have it publicly available, include this license, provide sufficient credit to the original authors of the project (IE: DaPorkchop_), as well as provide a link to the original project.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- */
-
-package net.daporkchop.lib.http.util.header;
-
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.experimental.Accessors;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.StringJoiner;
-
-/**
- * An HTTP header.
- *
- * @author DaPorkchop_
- */
-public interface Header {
-    /**
-     * @return the key of this header
-     */
-    String key();
-
-    /**
-     * @return the raw value of this header
-     */
-    String value();
-
-    /**
-     * @return a {@link List} containing all the values for this header
-     */
-    default List<String> valueList() {
-        return Arrays.asList(this.value().split(", "));
-    }
-
-    /**
-     * A simple implementation of {@link Header}.
-     *
-     * @author DaPorkchop_
-     */
-    @RequiredArgsConstructor
-    @Getter
-    @EqualsAndHashCode
-    @Accessors(fluent = true)
-    final class Default implements Header {
-        @NonNull
-        protected final String key;
-        @NonNull
-        protected final String value;
-
-        @Override
-        public String toString() {
-            return String.format("%s: %s", this.key, this.value);
-        }
-    }
-
-    /**
-     * A simple implementation of {@link Header} backed by a list of values.
-     *
-     * @author DaPorkchop_
-     */
-    @RequiredArgsConstructor
-    @Getter
-    @EqualsAndHashCode
-    @Accessors(fluent = true)
-    final class DefaultList implements Header {
-        @NonNull
-        protected final String key;
-        @NonNull
-        protected final List<String> valueList;
-
-        @Override
-        public String value() {
-            return this.valueList.stream().collect(() -> new StringJoiner(", "), StringJoiner::add, StringJoiner::merge).toString();
-        }
-
-        @Override
-        public String toString() {
-            return String.format("%s: %s", this.key, this.value());
-        }
-    }
-}
